go.text/language: added tag matching algorithm. This algorithm is not based on the CLDR algorithm. It incorporates some ideas from other implementations, but it is designed from scratch. Note that the IANA registry has been updated so this CL also adds some new language codes as well as add mappings for deprecated codes. maketables.go has also been modified to work around a bug that was introduced in the latest IANA update.
Note that the Match method of the Matcher interface returns an index of
the original Tag along with the Tag. Certain users of Matcher, such as
service like collation, need to associate data with each Tag.

Package collate is updated to use the new Matcher interface.
As the Matcher matches returns the index of the Tag as well, the tables
can be simplified to be an array instead of a map.

R=r, nigeltao, mpvl
CC=golang-dev, markdavis
https://codereview.appspot.com/13819047