go.tex/unicode/norm: added implementation of Transformer to norm.Form. One fundemental design decision was made to have the norm.Form type implement Transform directly, rather than having the user create a Transformer instance. The advantage of this approach is that it   1) results in a much nicer API (e.g. norm.NFC can be used as a      transformer as is).   2) the Transformer is stateless, thread-safe and reentrant, reducing      the possibility of errors.   3) is consistent with most other transformers. The clear disadvantage is that it is impossible to reuse a reorderBuffer between calls. The cost of initialization can be amortized when normalizing large blocks, but can be prohibitive for small strings (can probably get its size down to 500 bytes, but still).
However, in theory it is possible to skip the use of the reorderBuffer in
the fast majority of cases. 99.98 of HTML page content (excluding markup)
is in NFC (http://www.macchiato.com/unicode/nfc-faq). In most cases, NFC
can be converted to NFD without reordering (if it is in FCD form, which
can easily be detected). This means that, using the same techniques as for
norm.Iter, most conversions can be done without using a reorderBuffer at
all. If it is really necessary, it is probably possible to get rid of the
reorderBuffer altogether or to have an alternative that covers 99% of the
remaining cases.

As we reckon that the creation of a reorderBuffer can be avoided in most
cases, we opt for the better API. Note that this approach will be more
efficient if the creation of a reorderBuffer can be avoided.

The current implementation doesn't do any of the optimizations described
above. It only uses quickSpan to quickly cover most normalized content.
This will not avoid using a reorderBuffer when converting between different
forms, though.

Also:
- quickSpan has been modified to allow for incremental procesing (passing
atEOF) and returning whether any violating runes have been found (returning
a position smaller than the input length no longer signals this if !atEOF).
QuickSpan and QuickSpanString now no longer require a reorderBuffer.

- Added benchmarks to compare different normalization methods. Also added
benchmarks to compare these implementations against running ToLower on
the same strings.

All in all this CL raises some issues that will have to be addressed in
follow-up CLs:
 - Optimization of Transform method.
 - QuickSpan* probably needs to be adjusted to allow for incremental
   calls (like the internal version).
 - Unifying some of the implementations should be considered.

R=r
CC=golang-dev, nigeltao
https://codereview.appspot.com/23460044