go.text/language: added Parent() method that returns the CLDR parent of a tag.
Also added tie-breaker rule to Matcher to prefer matches for which there is
a parent relation over those where this is not the case. Note that this is a
low-priority rule as the region distance is a better measure (close regions
are likely to share the same parent; this is not true for English, but this
is already special-cased in a different manner). Also script and region
equality are stronger measures than the parent relationship.

The Parent relationship may change between different versions of CLDR
(albeit usually only slightly). Added a Version constant to identify CLDR
version on which the package is based.

LGTM=r
R=r
CC=golang-codereviews, markdavis
https://codereview.appspot.com/74210044