go.text/display: added display package, which provides an API for the localeDisplayNames data of CLDR.
The tables are quite large (1.6M), as this is a huge data set. The
design of this package is such that unused tables can be dropped
by the linker as much as possible. For example, just supporting
English can get you down to 6K.

The SelfNamer supports the case that users only care to provide an
option to select a locale, but not about displaying any futher names.
The SelfNamer provides names for a large number of languages in the
native language needing about 3K of data.

Note the somewhat strange naming of "U.S. English" vs "British English".
This is "fixed" in the current CLDR version which I will update this
package to soon.

It might be a possiblity to provide various versions of this package
with different levels of coverage or ones that (partly) uses OS-provided
names. In that case it may be an idea to define the Namer interface in
package language, at least, or immediately define a subpackage for
larger data sets. However, I think that if we ever do this we can
retain the current API and still make it efficient. Also, the currently
provided mechanism to keep table sizes under control might be sufficient.
Thoughts are welcome.

LGTM=r, nigeltao
R=r, nigeltao
CC=golang-codereviews
https://codereview.appspot.com/81410043