go.text/internal/ucd: support part indicaters (free-form lines starting with @).
An alternative API would be to pass the part and the comment as strings,
instead of passing the parser. There are a few small advantages to this:
- future proof in case this ever gets structured (very unlikely)
- if we ever support a line number in a parser, users might want to access it.
- users might not care about the comment, in which case it saves an alloc (doesn't matter)
- User could use Enum method to retrieve the value (useful in some cases).

But if having func(field, comment string) as a callback is considered to be nicer,
that is fine with me too.

Also fixes a bug where comments of parts weren't cleared for the next line.

LGTM=nigeltao
R=nigeltao
CC=golang-codereviews
https://codereview.appspot.com/135860043